use an event queue for keyboard and mouse packets
refactors how the VDUStreamProcessor handles keyboard and mouse events

introduces an `eventQueue`, which uses a custom thread-safe queue.  this queue is FIFO, and supports a `pushUnique` method to only push an entry it another of the same class is not already present.  this style of pushing is used for the keyboard and mouse events, in place of the previous boolean flags

this queue system will also allow us to add audio events to the queue handler at a later date, allowing this queue to be used for audio event handling including event callbacks in a manner that will be thread-safe

this uses C++17 features, so the build config has now been updated to use that