sprite based text cursor prep work
work in progress commit…  not working yet

In principle, this adjusts the text cursor handling to work with a text cursor sprite.  code that attempts to move the cursor has updates to move the sprite, if present.  code that may affect the appearance of the text cursor has been adjusted to call an “updateTextCursorBitmap” function, although that function is not yet completes

cursor flashing and visibility code has been updated.  the cursor now no longer needs to get temporarily hidden and re-shown during VDU command processing.

next step is to add support for a custom/separate text cursor sprite inside vdp-gl, and probably also support for making the mouse cursor a hardware sprite too.  this will affect the rest of the work here inside agon-vdp