fixes for scroll protect and paged mode
fixes issues around cursor movement with scroll protect, and how that interacts with paged mode and also holding down ctrl to introduce delays on newlines, and holding ctrl+shift to pause text output

additionally will now un-pause paged mode if “escape” is pressed.  unlike on Acorn sytems this may result in some pending output being printed to the screen.  given the Agon’s architecture this feels like a reasonable compromise

fundamentally, the changes are around better ensuring output gets paused when it should, and the cursor position is correctly

However… as of this commit whilst behaviour when an auto-newline is happening is the same as on an Acorn system when scroll protect is enabled, it differs when it’s disabled. an Acorn system will pause with the text cursor at the bottom right and the CRLF not yet done - whereas on the Agon we’ll have done the CRLF