fix tab, and text cursor screen positioning
tab had a minor logical error in how it was calculating positioning when the cursor behaviour is set to mean that axes are inverted.  as we’re subtracting based on the font character size, 1 needs to be added to the edge.  without this the position is off by 1 pixel, and a tab to the end column or row will overshoot by 1 pixel and be rejected

the logic that ensures that the text cursor remains on-screen, even when scroll protect might mean it’s just off-screen, has been updated to ensure it works correctly when cursor behaviour is set to invert and/or flip the axes